refactor dates helper import
