--- api/src/pdc/services/export/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "@/pdc/helpers/dates.helper.ts";
