+++ api/src/pdc/services/export/models/Campaign.ts
@@ -1,8 +1,8 @@
 import { ConfigInterfaceResolver } from "@/ilos/common/index.ts";
 import { get } from "@/lib/object/index.ts";
+import { toTzString } from "@/pdc/helpers/dates.helper.ts";
 import { Timezone } from "@/pdc/providers/validator/index.ts";
 import { SingleResultInterface as RawCampaignInterface } from "@/shared/policy/list.contract.ts";
-import { toTzString } from "../helpers/index.ts";
 
 export enum CampaignMode {
   Normal = "normal",
